+++ browser/BaseForm.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using System.ComponentModel;
 using System.Drawing;
 using System.Windows.Forms;
 
@@ -23,6 +24,66 @@ namespace MantisSharp.Browser
       this.ShowInTaskbar = false;
       this.StartPosition = FormStartPosition.CenterParent;
       this.FormBorderStyle = FormBorderStyle.FixedDialog;
+    }
+
+    #endregion
+
+    #region Properties
+
+    [DefaultValue(typeof(AutoScaleMode), "Font")]
+
+    public new AutoScaleMode AutoScaleMode
+    {
+      get { return base.AutoScaleMode; }
+      set { base.AutoScaleMode = value; }
+    }
+
+    [DefaultValue(typeof(FormBorderStyle), "FixedDialog")]
+
+    public new FormBorderStyle FormBorderStyle
+    {
+      get { return base.FormBorderStyle; }
+      set { base.FormBorderStyle = value; }
+    }
+
+    [DefaultValue(false)]
+
+    public new bool MaximizeBox
+    {
+      get { return base.MaximizeBox; }
+      set { base.MaximizeBox = value; }
+    }
+
+    [DefaultValue(false)]
+
+    public new bool MinimizeBox
+    {
+      get { return base.MinimizeBox; }
+      set { base.MinimizeBox = value; }
+    }
+
+    [DefaultValue(false)]
+
+    public new bool ShowIcon
+    {
+      get { return base.ShowIcon; }
+      set { base.ShowIcon = value; }
+    }
+
+    [DefaultValue(false)]
+
+    public new bool ShowInTaskbar
+    {
+      get { return base.ShowInTaskbar; }
+      set { base.ShowInTaskbar = value; }
+    }
+
+    [DefaultValue(typeof(FormStartPosition), "CenterParent")]
+
+    public new FormStartPosition StartPosition
+    {
+      get { return base.StartPosition; }
+      set { base.StartPosition = value; }
     }
 
     #endregion