+++ CHANGELOG.md
@@ -1,9 +1,11 @@
 Change Log
 ==========
 
-## 0.0.3
+## 0.3.0
 ### Added
 * Added `MantisClient.DeleteIssue(int)`
+* Added `MantisClient.GetConfigValues` with various overloads for returning configuration settings.
+* Updated Browser sample to include support for retrieving configuration settings
 
 ## 0.0.2